requirements.txt
Package            Version
------------------ ---------
certifi            2024.8.30
charset-normalizer 3.4.0
idna               3.10
pip                24.0
psycopg2-binary    2.9.10
python-dotenv      1.0.1
requests           2.32.3
urllib3            2.2.3
